[FIX] mail: prevent deletion of mail template attachments
Issue:
In recent 18.0 PRs (#186617 and #199922), new functionality was
added to the `MailComposer`. As a result, we now hit the
`mail/attachment/delete` controller when we delete an attachment
in the wizard due to the `unlink` call in `onFileRemove`.

However, this causes an unintended side effect with `mail.template`
records. When `_generate_template_attachments` is called, the
`attachment_ids` (ex. manually attached PDFs on the template to be
used rather than dynamically generated reports) use that same
`ir.attachment` record id, rather than making a copy. This makes
sense, since we don't want to explode the storage by making a copy
of these PDFs every time that `mail.template` is used.

But, since we `unlink` the attachment via the Composer (removing it
from the database) the `mail.template` is affected as well, resulting
in that `attachment_id` being missing from the template.

Steps to Reproduce:
1) Add an attachment to any `mail.template`.
2) Select that template via a `MailComposer`.
3) Delete the attachment that appears in the composer.
4) Notice the attachment is also removed from the db and template.

Solution:
We only want to remove the `ir.attachment` record from the database
if it is no longer going to be used, which isn't the case if we have
it set on the template. In `onFileRemove`, we can check to see if
the `fileId` to be removed matches any of the `attachment_ids` on
the template we are currently using, allowing us to determine if it is
safe to remove from the database completely or just from the
`MailComposer`.

+++ addons/mail/static/src/core/web/mail_composer_attachment_list.js
@@ -20,7 +20,7 @@ export class MailComposerAttachmentList extends Many2ManyBinaryField {
     async onFileRemove(fileId) {
         super.onFileRemove(fileId);
         const attachment = this.mailStore.Attachment.insert(fileId);
-        if (attachment) {
+        if (attachment && !this.props.record.data.template_id[1].includes(fileId)) {
             await this.attachmentUploadService.unlink(attachment);
         }
         this.env.fullComposerBus.trigger("ATTACHMENT_REMOVED", {
